requirements.txt
djangorestframework
django
pandas
markdown
django-filter
channels
django-cors-headers